restructured code around future statemachine view design
