+++ src/components/survey/defaultQuestions.ts
@@ -1,0 +1,14 @@
+export default [
+  {
+    fieldName: 'responsibility',
+    text: 'How often is this student responsible?',
+  },
+  {
+    fieldName: 'leadership',
+    text: 'How often does this student demonstrate leadership?',
+  },
+  {
+    fieldName: 'initiative',
+    text: 'How often does this student demonstrate initiative?',
+  },
+];
