--- src/pages/survey/defaultQuestions.ts
@@ -1,14 +1,0 @@
-export default [
-  {
-    fieldName: 'responsibility',
-    question: 'How often is this student responsible?',
-  },
-  {
-    fieldName: 'leadership',
-    question: 'How often does this student demonstrate leadership?',
-  },
-  {
-    fieldName: 'initiative',
-    question: 'How often does this student demonstrate initiative?',
-  },
-];
